Hide hair on pawns if hood is enabled
